Move Icon to src/components
